+++ src/pland/gui/LandManagerGUI.h
@@ -20,12 +20,19 @@ public:
     LDAPI static void confirmParentDelete(Player& player, SharedLand const& ptr); // 删除父领地
     LDAPI static void confirmMixDelete(Player& player, SharedLand const& ptr);    // 删除混合领地
 
-    LDAPI static void sendEditLandNameGUI(Player& player, SharedLand const& ptr);                 // 编辑领地名称
-    LDAPI static void sendEditLandDescGUI(Player& player, SharedLand const& ptr);                 // 编辑领地描述
-    LDAPI static void sendTransferLandGUI(Player& player, SharedLand const& ptr);                 // 转让领地
-    LDAPI static void _sendTransferLandToOfflinePlayerGUI(Player& player, SharedLand const& ptr); // 转让领地给离线玩家
-    LDAPI static void sendCreateSubLandConfirm(Player& player, SharedLand const& ptr);            // 创建子领地确认
-    LDAPI static void sendChangeRangeConfirm(Player& player, SharedLand const& ptr);              // 更改领地范围
+    LDAPI static void sendEditLandNameGUI(Player& player, SharedLand const& ptr);              // 编辑领地名称
+    LDAPI static void sendEditLandDescGUI(Player& player, SharedLand const& ptr);              // 编辑领地描述
+    LDAPI static void sendTransferLandGUI(Player& player, SharedLand const& ptr);              // 转让领地
+    LDAPI static void _sendTransferLandToOnlinePlayer(Player& player, SharedLand const& ptr);  // 转让领地给在线玩家
+    LDAPI static void _sendTransferLandToOfflinePlayer(Player& player, SharedLand const& ptr); // 转让领地给离线玩家
+    LDAPI static void _confirmTransferLand( // 确认转让领地
+        Player&           player,
+        SharedLand const& ptr,
+        mce::UUID         target,
+        std::string       displayName
+    );
+    LDAPI static void sendCreateSubLandConfirm(Player& player, SharedLand const& ptr); // 创建子领地确认
+    LDAPI static void sendChangeRangeConfirm(Player& player, SharedLand const& ptr);   // 更改领地范围
 
     LDAPI static void sendChangeMemberGUI(Player& player, SharedLand ptr);                     // 更改成员
     LDAPI static void _sendAddMemberGUI(Player& player, SharedLand ptr);                       // 添加在线成员